make sure we install the package only when it lives in the node_module folder

+++ scripts/install.js
@@ -2,28 +2,31 @@ const path = require('path');
 const fs = require('fs');
 const fse = require('fs-extra');
 
-// move sources
-fse.removeSync('js');
-fse.moveSync('dist/js','js');
-fse.removeSync('sass');
-fse.moveSync('src/sass','sass');
+if (process.env.PWD.match(/node_modules/)) {
+	// move sources
+	fse.removeSync('js');
+	fse.moveSync('dist/js','js');
+	fse.removeSync('sass');
+	fse.moveSync('src/sass','sass');
 
-// clean repo
-fse.removeSync('_toMigrate');
-fse.removeSync('dist');
-fse.removeSync('scripts');
-fse.removeSync('src');
-fse.removeSync('tests');
-fse.removeSync('.sass-cache');
-fse.removeSync('.babelrc');
-fse.removeSync('.changelogrc');
-fse.removeSync('.DS_Store');
-fse.removeSync('.editorconfig');
-fse.removeSync('.node-version');
-fse.removeSync('.travis.yml');
-fse.removeSync('favicon.ico');
-fse.removeSync('karma.conf.js');
-fse.removeSync('webpack.config.js');
+	// clean repo
+	fse.removeSync('_toMigrate');
+	fse.removeSync('dist');
+	fse.removeSync('scripts');
+	fse.removeSync('src');
+	fse.removeSync('tests');
+	fse.removeSync('.sass-cache');
+	fse.removeSync('.babelrc');
+	fse.removeSync('.changelogrc');
+	fse.removeSync('.DS_Store');
+	fse.removeSync('.editorconfig');
+	fse.removeSync('.node-version');
+	fse.removeSync('.travis.yml');
+	fse.removeSync('favicon.ico');
+	fse.removeSync('karma.conf.js');
+	fse.removeSync('webpack.config.js');
 
-// update sass files at root
-fs.writeFileSync('_index.scss', '@import "sass/sugar";');
+	// update sass files at root
+	fs.writeFileSync('_index.scss', '@import "sass/sugar";');
+}
+
